LayoutTests/fast/canvas/webgl/lose-context-on-timeout-expected.txt

CONSOLE MESSAGE: WebGL: context lost.
Checks that a GPU process timeout on a result producing WebGL call will lose the context.
NOTE: This only passes in the test harness because it requires Internals.

PASS gl.isContextLost() is false
PASS gl.getError() is gl.NO_ERROR
PASS color[0] is 77
PASS val is <= 998
PASS gl.isContextLost() is true
PASS gl.getError() is gl.CONTEXT_LOST_WEBGL
PASS successfullyParsed is true

TEST COMPLETE
